setup-mfa/requirements.txt
pyotp
qrcode
mysql-connector-python
Pillow